+++ src/Workspaces.Core/FindSymbols/SymbolFinderOptions.cs
@@ -7,12 +7,13 @@ namespace Roslynator.FindSymbols;
 internal class SymbolFinderOptions : SymbolFilterOptions
 {
     internal SymbolFinderOptions(
+        FileSystemFilter fileSystemFilter = null,
         VisibilityFilter visibility = VisibilityFilter.All,
         SymbolGroupFilter symbolGroups = SymbolGroupFilter.TypeOrMember,
         IEnumerable<SymbolFilterRule> rules = null,
         IEnumerable<AttributeFilterRule> attributeRules = null,
         bool ignoreGeneratedCode = false,
-        bool unusedOnly = false) : base(visibility, symbolGroups, rules, attributeRules)
+        bool unusedOnly = false) : base(fileSystemFilter, visibility, symbolGroups, rules, attributeRules)
     {
         IgnoreGeneratedCode = ignoreGeneratedCode;
         UnusedOnly = unusedOnly;