+++ src/Workspaces.Core/Rename/SymbolProvider.cs
@@ -10,17 +10,17 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
+using Microsoft.Extensions.FileSystemGlobbing;
 
 namespace Roslynator.Rename;
 
+#pragma warning disable RS1001, RS1022
+
 internal class SymbolProvider
 {
-    public SymbolProvider(bool includeGeneratedCode = false)
-    {
-        IncludeGeneratedCode = includeGeneratedCode;
-    }
+    public bool IncludeGeneratedCode { get; init; }
 
-    public bool IncludeGeneratedCode { get; }
+    public Matcher FileSystemMatcher { get; init; }
 
     public async Task<IEnumerable<ISymbol>> GetSymbolsAsync(
         Project project,
@@ -46,9 +46,13 @@ internal class SymbolProvider
             _ => throw new InvalidOperationException(),
         };
 
-        var analyzer = new Analyzer(
-            symbolKinds,
-            IncludeGeneratedCode);
+        var analyzer = new Analyzer()
+        {
+            IncludeGeneratedCode = IncludeGeneratedCode,
+            FileSystemMatcher = FileSystemMatcher,
+        };
+
+        analyzer.SymbolKinds.AddRange(symbolKinds);
 
         var compilationWithAnalyzers = new CompilationWithAnalyzers(
             compilation,
@@ -60,7 +64,6 @@ internal class SymbolProvider
         return analyzer.Symbols;
     }
 
-    [SuppressMessage("MicrosoftCodeAnalysisCorrectness", "RS1001:Missing diagnostic analyzer attribute.")]
     private class Analyzer : DiagnosticAnalyzer
     {
         [SuppressMessage("MicrosoftCodeAnalysisReleaseTracking", "RS2008:Enable analyzer release tracking")]
@@ -77,19 +80,15 @@ internal class SymbolProvider
 
         private static readonly ImmutableArray<DiagnosticDescriptor> _supportedDiagnostics = ImmutableArray.Create(DiagnosticDescriptor);
 
-        public Analyzer(ImmutableArray<SymbolKind> symbolKinds, bool includeGeneratedCode = false)
-        {
-            SymbolKinds = symbolKinds;
-            IncludeGeneratedCode = includeGeneratedCode;
-        }
-
         public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => _supportedDiagnostics;
 
         public ConcurrentBag<ISymbol> Symbols { get; } = new();
 
-        public ImmutableArray<SymbolKind> SymbolKinds { get; }
+        public List<SymbolKind> SymbolKinds { get; } = new();
 
-        public bool IncludeGeneratedCode { get; }
+        public bool IncludeGeneratedCode { get; init; }
+
+        public Matcher FileSystemMatcher { get; init; }
 
         public override void Initialize(AnalysisContext context)
         {
@@ -99,7 +98,7 @@ internal class SymbolProvider
                 ? GeneratedCodeAnalysisFlags.Analyze
                 : GeneratedCodeAnalysisFlags.None);
 
-            context.RegisterSymbolAction(f => AnalyzeSymbol(f), SymbolKinds);
+            context.RegisterSymbolAction(f => AnalyzeSymbol(f), SymbolKinds.ToArray());
         }
 
         private void AnalyzeSymbol(SymbolAnalysisContext context)
@@ -116,7 +115,7 @@ internal class SymbolProvider
                 case SymbolKind.NamedType:
                 case SymbolKind.Property:
                     {
-                        Symbols.Add(symbol);
+                        AddSymbol(symbol);
                         break;
                     }
                 case SymbolKind.Method:
@@ -130,7 +129,7 @@ internal class SymbolProvider
                             case MethodKind.UserDefinedOperator:
                             case MethodKind.Conversion:
                                 {
-                                    Symbols.Add(methodSymbol);
+                                    AddSymbol(methodSymbol);
                                     break;
                                 }
                         }
@@ -143,6 +142,12 @@ internal class SymbolProvider
                         break;
                     }
             }
+        }
+
+        private void AddSymbol(ISymbol symbol)
+        {
+            if (FileSystemMatcher?.IsMatch(symbol) != false)
+                Symbols.Add(symbol);
         }
     }
 }